Module Mirage_crypto_entropy

type t

The type of the entropy device.

type source = [

| `Timer

| `Rdseed

| `Rdrand

]

A polymorphic variant of entropy sources.

val pp_source : Stdlib.Format.formatter -> source -> unit

pp_source ppf source pretty-prints source on ppf.

val sources : unit -> source list

sources () is a list of supported entropy sources on your platform.

val initialize : ?⁠g:'a -> (module Mirage_crypto_rng.Generator with type g = 'a) -> t Lwt.t

initialize ~g rng_module sets the default generator to the rng_module and sets up periodic entropy feeding for that rng. This function raises if it is called a second time.